Shared data with workspaces, packages, lerna
