+++ client/package.json
@@ -1,12 +1,12 @@
 {
-  "name": "Gunn-WATT",
+  "name": "client",
   "version": "0.1.0",
   "private": true,
   "dependencies": {
     "@headlessui/react": "^1.6.4",
     "firebase": "^9.8.2",
     "he": "^1.2.0",
-    "luxon": "^2.4.0",
+    "luxon": "^3.0.1",
     "react": "^18.1.0",
     "react-color": "^2.19.3",
     "react-dom": "^18.1.0",
