+++ lerna.json
@@ -1,0 +1,11 @@
+{
+  "packages": [
+    "client",
+    "functions",
+    "scripts",
+    "shared"
+  ],
+  "useNx": false,
+  "version": "0.0.0",
+  "useWorkspaces": true
+}
