+++ package.json
@@ -1,0 +1,15 @@
+{
+  "name": "Gunn-WATT",
+  "version": "0.1.0",
+  "private": true,
+  "workspaces": [
+    "client",
+    "functions",
+    "scripts",
+    "shared"
+  ],
+  "devDependencies": {
+    "lerna": "^5.1.8",
+    "typescript": "^4.7.4"
+  }
+}
