+++ scripts/package.json
@@ -10,14 +10,16 @@
   },
   "license": "MIT",
   "dependencies": {
-    "@types/ical": "^0.8.0",
-    "@types/node": "^17.0.41",
-    "@types/node-fetch": "^2.5.12",
     "chalk": "^4.1.2",
     "firebase-admin": "^10.2.0",
     "ical": "^0.8.0",
     "node-fetch": "^2.6.7",
-    "ts-node": "^10.8.1",
-    "typescript": "^4.7.3"
+    "ts-node": "^10.8.1"
+  },
+  "devDependencies": {
+    "@types/ical": "^0.8.0",
+    "@types/node": "^18.0.6",
+    "@types/node-fetch": "^2.6.2",
+    "typescript": "^4.7.4"
   }
 }
