+++ shared/package.json
@@ -1,0 +1,8 @@
+{
+  "name": "shared",
+  "version": "0.1.0",
+  "description": "Shared utilities between the client, functions, and scripts.",
+  "devDependencies": {
+    "typescript": "^4.7.4"
+  }
+}
